+++ Tests/BugsnagBreadcrumbsTest.m
@@ -8,6 +8,7 @@
 
 #import "Bugsnag.h"
 #import "BugsnagClient.h"
+#import "BugsnagClientInternal.h"
 #import "BugsnagBreadcrumb.h"
 #import "BugsnagBreadcrumbs.h"
 #import "BugsnagTestConstants.h"
@@ -86,13 +87,6 @@ BSGBreadcrumbType BSGBreadcrumbTypeFromString(NSString *value);
 - (void)testMaxMaxBreadcrumbs {
     self.crumbs.capacity = 250;
     XCTAssertEqual(100, self.crumbs.capacity);
-}
-
-- (void)testClearBreadcrumbs {
-    [self.crumbs clearBreadcrumbs];
-    awaitBreadcrumbSync(self.crumbs);
-    XCTAssertTrue(self.crumbs.count == 0);
-    XCTAssertNil(self.crumbs[0]);
 }
 
 - (void)testEmptyCapacity {
@@ -176,8 +170,7 @@ BSGBreadcrumbType BSGBreadcrumbTypeFromString(NSString *value);
 }
 
 - (void)testDefaultDiscardByType {
-    [self.crumbs clearBreadcrumbs];
-    awaitBreadcrumbSync(self.crumbs);
+    self.crumbs = [BugsnagBreadcrumbs new];
     [self.crumbs addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull crumb) {
         crumb.type = BSGBreadcrumbTypeState;
         crumb.message = @"state";
@@ -223,8 +216,7 @@ BSGBreadcrumbType BSGBreadcrumbTypeFromString(NSString *value);
 }
 
 - (void)testAlwaysAllowManual {
-    [self.crumbs clearBreadcrumbs];
-    awaitBreadcrumbSync(self.crumbs);
+    self.crumbs = [BugsnagBreadcrumbs new];
     self.crumbs.enabledBreadcrumbTypes = 0;
     [self.crumbs addBreadcrumb:@"this is a test"];
     awaitBreadcrumbSync(self.crumbs);
@@ -239,8 +231,7 @@ BSGBreadcrumbType BSGBreadcrumbTypeFromString(NSString *value);
  * private and assumes filtering is already configured.
  */
 - (void)testDiscardByTypeDoesNotApply {
-    [self.crumbs clearBreadcrumbs];
-    awaitBreadcrumbSync(self.crumbs);
+    self.crumbs = [BugsnagBreadcrumbs new];
     self.crumbs.enabledBreadcrumbTypes = BSGEnabledBreadcrumbTypeProcess;
     // Don't discard this
     [self.crumbs addBreadcrumbWithBlock:^(BugsnagBreadcrumb *_Nonnull crumb) {